requirements.txt
Django==2.2.7
django-tinymce4-lite==1.7.5
jsmin
pytz==2019.3
sqlparse==0.3.0
